# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "习惯打卡系统 API",
    "version": "1.0.0",
    "description": "一个习惯追踪和打卡系统,支持用户注册、习惯管理、打卡记录和AI助手功能"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "开发服务器"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Checkin": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "habitId": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "note": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/checkins": {
      "get": {
        "summary": "获取打卡记录",
        "tags": ["打卡"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}$"
            },
            "description": "查询月份(格式:YYYY-MM)"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取打卡记录",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Checkin"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "删除打卡记录",
        "tags": ["打卡"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "打卡记录ID"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        }
      }
    },
...
  },
...
}